Move process class into stage
